requirements.txt
# CoffeeBliss FastAPI backend requirements
fastapi==0.110.2
uvicorn[standard]==0.29.0
SQLAlchemy[asyncio]==2.0.29
asyncpg==0.29.0
python-jose==3.3.0
bcrypt==4.1.2
python-dotenv==1.0.1
pydantic==2.7.1
